+++ bin/tests/spiders_dot_py_tests.py
@@ -41,14 +41,14 @@ class TestSpidersDotPy(unittest.TestCase):
           "cloudfeaster.samples.pypi_spider.PyPISpider": {
             "url": "http://pypi-ranking.info/alltime",
             "factor_display_names": {},
-            "ttl": 60,
+            "ttl_in_seconds": 60,
             "max_concurrency": 3,
             "factor_display_order": []
           },
           "cloudfeaster.samples.bank_of_canada_daily_exchange_rates.BankOfCanadaDailyExchangeRatesSpider": {
             "url": "http://www.bankofcanada.ca/rates/exchange/daily-exchange-rates/",
             "factor_display_names": {},
-            "ttl": 60,
+            "ttl_in_seconds": 60,
             "max_concurrency": 3,
             "factor_display_order": []
           }